Extract components into separate files
