+++ desktop/apps/categories3/components/Gene.js
@@ -1,0 +1,11 @@
+import React from 'react'
+
+const Gene = ({ id, name }) => {
+  return (
+    <li>
+      {name}
+    </li>
+  )
+}
+
+export default Gene
