+++ desktop/apps/categories3/components/GeneFamilies.js
@@ -1,0 +1,14 @@
+import React from 'react'
+import GeneFamily from './GeneFamily'
+
+const GeneFamilies = ({ geneFamilies }) => {
+  return (
+    <div>
+      {geneFamilies.map(geneFamily =>
+        <GeneFamily key={geneFamily.id} {...geneFamily} />
+      )}
+    </div>
+  )
+}
+
+export default GeneFamilies
