+++ desktop/apps/categories3/components/GeneFamily.js
@@ -1,0 +1,17 @@
+import React from 'react'
+import Gene from './Gene'
+
+const GeneFamily = ({ id, name, genes }) => {
+  return (
+    <div>
+      <h2>
+        {name}
+      </h2>
+      <ul>
+        {genes.map(gene => <Gene key={gene.id} {...gene} />)}
+      </ul>
+    </div>
+  )
+}
+
+export default GeneFamily
