+++ desktop/apps/categories3/components/GeneFamilyNav.js
@@ -1,0 +1,11 @@
+import React from 'react'
+
+const GeneFamilyNav = ({ geneFamilies }) => {
+  return (
+    <ul>
+      {geneFamilies.map(geneFamily => <li key={geneFamily.id}>TK</li>)}
+    </ul>
+  )
+}
+
+export default GeneFamilyNav
