+++ desktop/apps/categories3/components/TAGPIntro.js
@@ -1,0 +1,17 @@
+import React from 'react'
+
+const TAGPIntro = () => {
+  return (
+    <div>
+      <h1>The Art Genome Project</h1>
+      <p>
+        Lorem ipsum dolor sit amet consectetur adipisicing elit. Sit, quidem
+        facere doloribus quisquam ducimus alias, consequuntur officia modi
+        dolores ipsam distinctio harum facilis, illum fugit ea? Quo illum esse
+        et!
+      </p>
+    </div>
+  )
+}
+
+export default TAGPIntro
